move tests to root/test/
